arquitetura organizada da pagina e components
