+++ pages/forex-page.py
@@ -3,81 +3,13 @@ import pandas as pd
 import numpy as np
 import dash_bootstrap_components as dbc
 import plotly.graph_objects as go
-from plotly.subplots import make_subplots
-from dash import Input, Output, html, dcc, callback
+from dash import dcc
 from classes.forex_coin import ForexCoin
 from classes.indicator import Indicator
+from components.graph import GraphComponent
 
-# Registra a página
-dash.register_page(__name__, path='/forex-page')
 
-# Layout da página
-layout = dbc.Container(
-    [
-        dbc.Row(
-            [
-                # GRAPH COIN
-                dcc.Graph(id='graph-forex'),
-                dcc.Interval(
-                    id='interval-graph-forex',
-                    interval=5*60*1000, # 5 min
-                    n_intervals=0
-                ),
-            ]
-        ),
-        
-    ], 
-    fluid=True
-)
-
-config_layout = lambda name: {
-    'title':dict(
-        text=name,
-        font=dict(size=20, color='white')  # Cor e tamanho do título
-    ),
-    'xaxis':dict(
-        title='',  # Título do eixo x
-        title_font=dict(size=14, color='white'),  # Cor e tamanho do título do eixo x
-        type='date',  # Define o tipo de dado como data
-        tickformat='%d-%m-%Y %H:%M',  # Formato dos ticks para o eixo de tempo
-        tickfont=dict(color='lightgray'),  # Cor das marcações dos ticks
-        tickangle=45,  # Ângulo dos ticks para melhor visualização
-        gridcolor='gray',  # Cor das linhas de grade verticais
-        showline=True,  # Mostra a linha do eixo x
-        linecolor='white',  # Cor da linha do eixo x
-    ),
-    'yaxis':dict(
-        title='',  # Título do eixo y
-        title_font=dict(size=14, color='white'),  # Cor e tamanho do título do eixo y
-        tickfont=dict(color='lightgray'),  # Cor das marcações dos ticks no eixo y
-        gridcolor='gray',  # Cor das linhas de grade horizontais
-        showline=True,  # Mostra a linha do eixo y
-        linecolor='white',  # Cor da linha do eixo y
-    ),
-    'template':'plotly_dark',  # Tema escuro
-    'height':900,
-    'legend':dict(
-        orientation='h',
-        bgcolor='rgba(0,0,0,0.5)',
-        bordercolor='white',
-        borderwidth=1,
-        font=dict(
-            size=12,
-            color='white'
-        )
-    )
-}
-
-@callback(
-    Output('graph-forex', 'figure'),
-    Input('interval-graph-forex', 'n_intervals')
-)
-def update_graph_forex(n):
-    
-    coin = ForexCoin('EURUSD')
-    coin.path = './data/forex/15m/EURUSD.csv'
-
-    df = coin.get_dataframe()
+def calculate_indicators(df: pd.DataFrame) -> pd.DataFrame:
 
     df['Percentual'] = ((df['Close'] - df['Open']) / df['Open']) * 100
     df['MP'] = Indicator.mean(df['Percentual'], 30) * 10
@@ -97,10 +29,17 @@ def update_graph_forex(n):
     test = Indicator.setup_test(df)
 
     df['Gains'] = test[test['Operation'] == 'Gain']['EMA_9'] * 1.001
-    
 
-    # Criando subplots
-    fig = make_subplots(rows=2, cols=1, shared_xaxes=True, vertical_spacing=0.1, row_heights=[.9, .1])
+    return df
+
+
+def update_graph_forex(fig:go.Figure):
+    
+    coin = ForexCoin('EURUSD')
+    coin.path = './data/forex/15m/EURUSD.csv'
+
+    df = coin.get_dataframe()
+    df = calculate_indicators(df)
 
     candlestick = lambda name, df: go.Candlestick( x=df.index, open=df['Open'], high=df['High'], low=df['Low'], close=df['Close'], name='Candlestick')
     line = lambda name, color, dfy: go.Scatter( x=df.index, y=dfy, mode='lines', name=name, line=dict(color=color), visible='legendonly')
@@ -122,13 +61,32 @@ def update_graph_forex(n):
     fig.add_trace(marker(f'SETUP 9.1 BUY ({n_setup_9_1_buy})', '#fff', df['_9_1_Buy']), row=1, col=1)
     fig.add_trace(marker(f'SETUP 9.1 SELL ({n_setup_9_1_sell})', '#fff', df['_9_1_Sell']), row=1, col=1)
 
-    fig.add_trace(marker(f'GAIN', '#0f0', df['Gains']), row=1, col=1)
-
-
-
-
-    fig.update_layout(**config_layout(coin.name))
-
-
+    percentage_gain = df['Percentage_Gain'].iloc[0]
+    fig.add_trace(marker(f'GAIN ({percentage_gain:.0f}%)', '#0f0', df['Gains']), row=1, col=1)
 
     return fig
+
+
+
+
+# Registra a página
+dash.register_page(__name__, path='/forex-page')
+
+graph_component = GraphComponent()
+
+# Layout da página
+layout = dbc.Container(
+    [
+        dbc.Row(
+            [
+                # GRAPH COIN
+                graph_component.graph
+            ]
+        ),
+        
+    ], 
+    fluid=True
+)
+
+graph_component.init_callback(dash.get_app())
+graph_component.fig = update_graph_forex(graph_component.fig)
